requirements.txt
autopep8==2.0.2
bidict==0.22.1
bpyutils==0.5.8
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pycodestyle==2.10.0
python-engineio==4.4.0
python-socketio==5.8.0
RPi.GPIO==0.7.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
